[refactor] #133 Version Live, Test 분리
